fix: Rename `SupportedDevices` -> `SupportedDevicesTable`
