+++ src/lib/ui/SupportedDevicesTable/SupportedDevicesContent.tsx
@@ -3,8 +3,8 @@ import axios from 'axios'
 import { useCallback, useEffect, useState } from 'react'
 
 import { Button } from 'lib/ui/Button.js'
-import { SupportedDevicesFilterArea } from 'lib/ui/SupportedDevices/SupportedDevicesFilterArea.js'
-import type { DeviceModel, Filters } from 'lib/ui/SupportedDevices/types.js'
+import { SupportedDevicesFilterArea } from 'lib/ui/SupportedDevicesTable/SupportedDevicesFilterArea.js'
+import type { DeviceModel, Filters } from 'lib/ui/SupportedDevicesTable/types.js'
 
 import { SupportedDeviceRow } from './SupportedDeviceRow.js'
 import { SupportedDevicesHeader } from './SupportedDevicesHeader.js'
