+++ src/lib/ui/SupportedDevicesTable/types.ts
