+++ src/lib/ui/index.ts
@@ -2,4 +2,4 @@ export * from './AccessCodeDetails/AccessCodeDetails.js'
 export * from './AccessCodeTable/AccessCodeTable.js'
 export * from './DeviceDetails/DeviceDetails.js'
 export * from './DeviceTable/DeviceTable.js'
-export * from './SupportedDevices/SupportedDevices.js'
+export * from './SupportedDevicesTable/SupportedDevicesTable.js'
